18th oct HW

+++ 13th OCT 2023/HW_Task2_1.py
@@ -1,7 +1,7 @@
 # Write a Python program to calculate the area of a circle given its radius using the formula
 # area=π×r^2 ( Take pie as 3.14)
 
-rad = int(input('Enter the desired radius:\n'))
+rad = float(input('Enter the desired radius:\n'))
 pi = 3.1416
 area = pi * (rad ** 2)
 print('The area of the circle with radius ' + str(rad) + ' is ' + str(area))
